refactor(shadowPass): update `layer` and viewport dimensions to use `int32_t` for type consistency

+++ src/rendering/renderPasses/shadowPass/directionalShadowPass.cpp
@@ -47,7 +47,7 @@ void DirectionalShadowPass::render(const RenderContext& ctx, const glm::vec4& di
 	glClear(GL_DEPTH_BUFFER_BIT);
 	glEnable(GL_DEPTH_TEST);
 	glCullFace(GL_FRONT);
-	glViewport(0, 0, ctx.shadow.width, ctx.shadow.height);
+	glViewport(0, 0, static_cast<int32_t>(ctx.shadow.width), static_cast<int32_t>(ctx.shadow.height));
 
 	for (const auto& [entity, material, shader, mesh]: ctx.renderQueue->shadowingObjects) {
 		RenderCommon::setupTransform(*entity, *mDepthShader);
